+++ Modules/Microsoft365DSC/DSCResources/MSFT_IntuneDeviceManagementAndroidDeviceOwnerEnrollmentProfile/MSFT_IntuneDeviceManagementAndroidDeviceOwnerEnrollmentProfile.psm1
@@ -477,7 +477,10 @@ function Export-TargetResource
     try
     {
         $Script:ExportMode = $true
-        [array] $Script:exportedInstances = Get-MgBetaDeviceManagementAndroidDeviceOwnerEnrollmentProfile -ErrorAction Stop
+        # Exclude profiles with Microsoft internal enrollment mode (EnrollmentMode 5) from export
+        # as it cannot be managed. Example is "Default enrollment profile for personally-owned work profile devices"
+        [array] $Script:exportedInstances = Get-MgBetaDeviceManagementAndroidDeviceOwnerEnrollmentProfile `
+            -ErrorAction Stop | Where-Object EnrollmentMode -NE 5
 
         $i = 1
         $dscContent = ''